requirements.txt
pyyaml>=6.0
jinja2>=3.1
pylatexenc>=2.10
jsonschema>=4.20
